Refactor eslint-parser to return promise from execute
