+++ packages/plugin-ember-octane/src/hooks/register-tasks.ts
@@ -1,5 +1,5 @@
 import { Hook } from '@oclif/config';
-import { OctaneMigrationStatusTask } from '../tasks';
+import OctaneMigrationStatusTask from '../tasks/octane-migration-status-task';
 
 // TODO: Determine correct type for options
 const hook: Hook<'register-tasks'> = async function ({ cliArguments, parsers, tasks }: any) {
