--- packages/plugin-ember-octane/src/results/index.ts
@@ -1,1 +1,0 @@
-export { default as OctaneMigrationStatusTaskResult } from './octane-migration-status-task-result';
