--- packages/plugin-ember-octane/src/tasks/index.ts
@@ -1,1 +1,0 @@
-export { default as OctaneMigrationStatusTask } from './octane-migration-status-task';
